pyproject.toml
[project]
name = "strands-handson"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "boto3==1.40.21",
    "feedparser>=6.0.12",
    "langchain==0.3.27",
    "langchain-aws==0.2.31",
    "langchain-mcp-adapters==0.1.9",
    "langchain-tavily==0.2.11",
    "langgraph==0.6.6",
    "pydantic==2.11.7",
    "python-dotenv>=1.2.1",
    "strands-agents>=1.21.0",
    "streamlit==1.49.1",
]
